More conversion to doubles Add a remove function for the compound renderer

+++ src/include/enesim_matrix.h
@@ -82,7 +82,7 @@ EAPI void enesim_matrix_divide(Enesim_Matrix *m, float scalar);
 EAPI void enesim_matrix_inverse(Enesim_Matrix *m, Enesim_Matrix *m2);
 EAPI void enesim_matrix_adjoint(Enesim_Matrix *m, Enesim_Matrix *a);
 
-EAPI void enesim_matrix_point_transform(Enesim_Matrix *m, float x, float y, float *xr, float *yr);
+EAPI void enesim_matrix_point_transform(Enesim_Matrix *m, double x, double y, double *xr, double *yr);
 EAPI void enesim_matrix_eina_rectangle_transform(Enesim_Matrix *m, Eina_Rectangle *r, Enesim_Quad *q);
 EAPI void enesim_matrix_rectangle_transform(Enesim_Matrix *m, Enesim_Rectangle *r, Enesim_Quad *q);
